chore: move testss for BindMounts and VolumeMounts to its test file
